Improving services and configuration (WIP)

+++ partnerbackendtest/src/main/java/pruebas/partner/model/Partner.java
@@ -7,15 +7,15 @@ public class Partner {
 
     private String name;
     private String status;
-    private Double partner_key;
-    private Double partner_id;
+    private Long partner_key;
+    private Long partner_id;
 
     public static Partner create(String name, String key, String status, String id) {
         Partner partner = new Partner();
         partner.name = name;
         partner.status = status;
-        partner.partner_key = new Double(key);
-        partner.partner_id = new Double(id);
+        partner.partner_key = new Long(key);
+        partner.partner_id = new Long(id);
         return partner;
     }
